arm/coproc: Add vGPIO coprocessor
Virtual GPIO (vGPIO) - virtual module which allows to set the order in
which interrupts coming from different domains will be processed.

Implement functional in which reading, writing and irq injection occurs
in according to rules, described in per-domain vcoproc configuration
(every domain has access to set of pins, assigned in the correspond
structure).

Configuration is described in dom.cfg file in coproc section. The pins
assigned to the domain configured with string "pins=" followed with a
32-bit hexadecimal mask. Where bit set to '1' means the correspond pin is
assigned to the domain, '0' - means the pin is not assigned to the domain.

Signed-off-by: Anastasiia Lukianenko <anastasiia_lukianenko@epam.com>
Reviewed-by: Oleksandr Andrushchenko <oleksandr_andrushchenko@epam.com>
Reviewed-by: Andrii Anisov <andrii_anisov@epam.com>