arm/coproc: Power on relevant SCU resource for vgpio
Find SCU's resource ID for the virtual GPIO and power it on.
Prepare infrastructure for dynamical power management, e.g.
power on on the very first vcoproc instance and power off when
the last is gone.

The relevant SCU's resource ID is defined in Xen's device tree in
"power-domains" property:

	gpio4: gpio@5d0c0000 {
		[...snip...]
                power-domains = <&pd_lsio_gpio4>;
		[...snip...]
	};

	pd_lsio_gpio4: PD_LSIO_GPIO_4 {
		[...snip...]
		reg = <SC_R_GPIO_4>;
		[...snip...]
	};

Please note, that "power-domains" property must be removed from
the relevant guest's device tree nodes.

Signed-off-by: Oleksandr Andrushchenko <oleksandr_andrushchenko@epam.com>