domain_build: Ignore disabled device-tree nodes
Currently, XEN on device tree parsing ignores the disabled status of
nodes. This leads to getting disabled nodes iomems and irqs are
marked as available and could be passed to guest domains. In such a
way devices disabled by firmware could be accidentally made available
for guest domains. To avoid such a situation, skip disabled nodes
from parsing and adopting.

Signed-off-by: Andrii Anisov <andrii_anisov@epam.com>
Reviewed-by: Anastasiia Lukianenko <anastasiia_lukianenko@epam.com>
Reviewed-by: Oleksandr Andrushchenko <oleksandr_andrushchenko@epam.com>