imx8qm: Skip imx8qm-pm node during passthrough device processing
In case of processing imx8qm-pm node different domains will be able to
control the same resources. This behavior is unacceptable, because
destroying one domain will result errors in all domains with the common
resources and domains will be able to manage resources which are not
intended to them.
So imx8qm-pm node must be skipped to avoid undefined failures and security
issues.

Signed-off-by: Anastasiia Lukianenko <anastasiia_lukianenko@epam.com>
Reviewed-by: Oleksandr Andrushchenko <oleksandr_andrushchenko@epam.com>
Tested-by: Oleksandr Andrushchenko <oleksandr_andrushchenko@epam.com>
Reviewed-by: Andrii Anisov <andrii_anisov@epam.com>