coproc: skip route irq to guest already routed to coproc
When XEN creates coproc, it routes its irq with the appropriate number.
We need to provide domu.cfg with a list of irqs to use irqs on the guest domain.
During the creation of the guest domain,
it will fail due to re-routing of this IRQ number.
To avoid double routing of the IRQ to the guest domain,
we skip routing it with a warning message.

Signed-off-by: Vasyl Gryshchenko <vasyl_gryshchenko@epam.com>

Reviewed-by: Sergiy Kibrik <sergiy_kibrik@epam.com>