p2m: set direct_nc_x type only for direct reserved memory mappings
p2m_mmio_direct_nc_x mapping type was added to support instructions
execution from the mapped memory. This is needed for the devices, such
as vpu_encoder and vpud_decoder which are using mapped memory to run
firmware. This type should be set only for the direct mapping, any other
types, such as ram or foreign should be set as is.

Signed-off-by: Oleksii Moisieiev <oleksii_moisieiev@epam.com>