vscmi: make the freq/power ratio monotonically decreasing
Some guest operating systems uses the power information transmitted
over SCMI to decide wheter or not an OPP shall be used or not.

Notably, the Linux kernel expects that frequency / power ratio to be
monotonically decreasing  - i.e., that OPPs operating at lower
frequencies are more efficient than the ones at higher frequencies.
If this is not true for a specific OPP, the kernel marks it as not
efficient, possibly avoiding to use it.

Since the current implementation in Xen uses the OPP index as power
value and the OPP frequencies are equally spaced, the aforementioned
ratio is equal for all the exposed OPPs and they are all marked as
inefficient except the first one.

Fix this behaviour using the square of the index as power value; while
this does not have any physical meaning, it ensures the correct trend
in OPP efficiency.

Signed-off-by: Francesco Valla <francesco.valla@mta.it>
Acked-by: Mykyta Poturai <mykyta_poturai@epam.com>
Acked-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>